AIContollerBasics
Attempts to measure the distance between player and enemy, some issue with shooting  - for this moment i dont know if enemies should posses weapon and whether I should repeat the code for shooting/add the weapon script to the enemy. Work in progress.
The changes in the weapon folder dont exist, again, I dont know how to delete a comit in some part, same with the one with a messed up crab folder, don't laugh at me, this is serious stuff really